+++ moshi-ir/moshi-compiler-plugin/src/main/kotlin/dev/zacsweers/moshix/ir/compiler/MoshiComponentRegistrar.kt
@@ -4,9 +4,7 @@ package dev.zacsweers.moshix.ir.compiler
 
 import com.google.auto.service.AutoService
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
-import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
-import org.jetbrains.kotlin.config.CommonConfigurationKeys
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.name.ClassId
 
@@ -19,15 +17,12 @@ public class MoshiComponentRegistrar : CompilerPluginRegistrar() {
 
   override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
     if (configuration[KEY_ENABLED] == false) return
-    val debug = configuration[KEY_DEBUG] == true
     val enableSealed = configuration[KEY_ENABLE_SEALED] == true
 
-    val messageCollector =
-      configuration.get(CommonConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
     val fqGeneratedAnnotation = configuration[KEY_GENERATED_ANNOTATION]?.let(ClassId::fromString)
 
     IrGenerationExtension.registerExtension(
-      MoshiIrGenerationExtension(messageCollector, fqGeneratedAnnotation, enableSealed, debug)
+      MoshiIrGenerationExtension(fqGeneratedAnnotation, enableSealed)
     )
   }
 }